requirements.txt
Django==2.2.1
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-js-asset==1.2.2
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
